requirements.txt
asgiref==3.9.1
Django==4.2.23
PyMySQL==1.1.2
sqlparse==0.5.3
Pillow==10.4.0